updates to api e2e test format and comments for future tests
